Extract shader instance editor properties panel to separate component.
